+++ examples/basic-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # basic-react-example
 
+## 0.1.1
+
+### Patch Changes
+
+- Add support for external phone number hook
+- Updated dependencies
+  - @mecaptcha/verify-react@0.2.1
+  - @mecaptcha/verify-sdk@0.2.1
+
 ## 0.1.0
 
 ### Minor Changes