+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mecaptcha/verify-sdk
 
+## 0.2.1
+
+### Patch Changes
+
+- Add support for external phone number hook
+
 ## 0.2.0
 
 ### Minor Changes